cleaned up folders and names
